fix: update mobile menu products slider for edge case
